requirements.txt
fastapi
uvicorn
python-telegram-bot
feedparser
openai
facebook-business
fuzzywuzzy
pillow
gspread
pandas
pytest
python-dotenv
beautifulsoup4
sentence-transformers
scikit-learn
facebook-scraper
selenium==4.21.0
puppeteer